Stubbed out opening Void with an empty buffer
Pair programming with @CraigStuntz at CodeMash. So awesome! He's
definitely ahead of me both on understanding of F# and in functional
thinking. Thanks, man! It was good!

My R# license timed out since I'm not connected to the license server,
so I haven't run the unit tests. We'll see if they pass... ;)