Hacked line height & placement to work on Windows
The sizing was not working right on Windows with the cleaner solution.
I'm going to pull this down on my Linux box and see if this hacked
version works there. Something inside me is wondering if there is a bug
in ETO. I am probably going to have problems with horizontal sizing,
too, once I get that in place. Right now I'm focused on getting this
project off the ground so I am not too concerned about the hack.