Began to move control from the UI to a controller
This involved the first reference to the Void.ViewModel assembly from
the Void assembly, which meant I had to recreate Void.ViewModel as a
portable assembly because otherwise it could not be added. This also
meant that I had to recreate View.Model as a portable assembly since it
is a dependency of Void.ViewModel.