Convert a core buffer to a view model buffer
In the process, finally gave up the uint16 crap and just fell back on
int. This seems to be a case where F#'s type system is not strong in a
helpful way. It is not easy, particularly when inter-oping with the rest
of .NET (though it is not easy period, according to what I know) to
impose meaningful constraints on numeric types the way you can on your
other types. Oh well!