Fixed redraw, and then fixed resultant flickering
However, the repaint does not always work correctly. For example,
minimizing and then maximizing the window leaves the screen black.

What we probably need to be doing is to create view model objects off
events coming out of the core, store them off, and trigger a localized
redraw. When the redraw occurs, we should capture the area that has been
invalidated and render that portion of the view model as drawing
objects, and draw those. Currently we are storing off drawing objects,
and that's not looking like it's going to work well, because either we
don't flush them, and then we get flickering; or we flush them, and then
we have nothing for legitimate re-paints. I have designed the view model
so this would be possible, but it is not until now that I have reached
full clarity on this.

Also, we need to test on Linux and see if this works cross-platform.