Trying to introduce the concept of windows into the core
I am trying to get things structured more the way they should be, with
the ViewModel being created off of events that come out of the core
which then triggers redraws which is then rendered into drawing objects
and draws them. I know like I have wondered from strictly trying a new
GUI library, but I want to make sure that it will work well in all
cases; but some of the cases I want to get working (such as the
refreshing) are going to need the architecture to be evolved more.

If the experiment fails, though, I will probably need to cherry-pick
some of the commits from this branch, since it has strayed a
bit/broadened its purpose.