Decided that parsing filepaths is editor logic
Decided that commands like ;edit should be parsed as taking raw input
because special things that can occur in the filepath such as # or #2
for buffer numbers or % for current path are really closely tied into
the editor and don't need to be baked into the language at a low level.

Note: the line commands that I have currently defined are still very
sketched-out and preliminary.