Introduce "text or hotkey" keyboard input
This will be needed for insert and command mode. To accomplish this, I
had to restructure the way in which input came down from the view. I
did this because I didn't want the ModeController to fictiously claim to
have changed the mode when it is really just hoping that something
downstream will actually change the way that input is interpreted.

Please note that some of what I introduced in this commit is a hack that
is intended to be solved by issue #2.