Removed windows and tabs from Void Core types
Realized after seeing an announcement about Neovim's Smart UI Protocol
(http://usevim.com/2015/06/24/neovim-smart-iu/) that windows and tabs
were really a ViewModel concern. This changes my understanding of the
view model somewhat. Whereas before I was thinking of it as a literal
model of what the user sees at a given moment, I am now beginning to see
it as a model of what viewports there are for buffers. Much more drastic
changes to come on this basis.