Finally got things working again after crazy overhaul!
I wanted to get ready for coding with @mjgpy3 this weekend, so I thought
I had better get stuff cleaned up on this branch. Of course, since it is
the middle of the night and I couldn't sleep, I also didn't control the
scope of my changes very well. Not that I've been doing a good job of
that anyway.

+ Restored the ;edit functionality.
+ Do not update service's state if noMessage was returned. This prevents
  needless updates. This may be a premature optimization.
+ Plug in the BufferList and WindowBufferMap services.
+ Limited support for expanding ~ to be $HOME in Filesystem service.
+ Updated Notifications to use new model of services.
+ Fixed some of the messaging by removing message from incorrect layers
  or removing overly fat message in favor of smaller, more meaningful
  ones.
+ Also touched up some of the documentation.