If nothing services a request, emit NoResponseToRequest<> message
This is a key piece of the request/response functionality, and
constitutes one of its strengths over command events; it allows services
to compensate if other services fail or do not exist, etc.

I decided to build out a little more functionality with changing the
command mode language to prove that this works... and found that it
DIDN'T work. Apparently `sequence = Seq.empty` is false even is the
sequence is empty; use Seq.isEmpty instead.

However the command mode language stuff in itself proved to be an
interesting use case for this. By use request/response semantics, the
command mode service is able to function without the command language
service even being subscribed to the bus. This is enabled precisely by
the fact that it is able to fall back on a default language is it does
not receive a response to its request.

Frankly, this makes me feel quite good about this request/response
stuff. It looks like it will allow us to break our services even
smaller. At first I just started introducing the concept of a command
language option in the command mode service, but I had to change
EVERYTHING about the service and its tests, and it was getting
complicated fast. Moreover, I noticed that command language switching
and managing the command buffer are orthogonal concerns, and it seemed
bad to lump them together even though once again I shocked myself by how
small I am breaking things down.

But I LOVE that if one service is nonexistent or malfunctioning, the
other can still function. It's the same thing we accomplished with the
command mode and command history services. Even though this is no a
distributed system, that has to be a solid way to build a system. Limit
the blast radius! It feels so good.

Which probably means there's something somewhere I'm not thinking of...

@mjgpy3 At this point I'm comfortable with saying, if you can
whole-heartedly approve of these things, merge away. By the way, I used
your message logging stuff when debugging one of the Void.Spec specs,
and it was great! Thanks again for that.