Decouple rendering of command bar from view model service
I went in to make my changes for scrolling, and once again the huge mass
that is the view model hindered me. I was somewhat at a loss for what to
do, so I thought I would start by finishing my decoupling of the command
bar, which I have already decoupled to a large degree, thinking that
might be instructive on how to proceed with scrolling in windows.

In Go terms, tesuji. I've played a move elsewhere.