Introduce units of measure for rows and columns
As I began to write my first set of specs for actual motion using hjkl,
I realized that not even I knew whether rows or lines where zero based
or one based, or whether it was one way in one place and another way in
another, or what the lay of the land was, and if even _I_, currently
almost the sole author, do not know that this early on, then it was
hopeless.

Enter F# units of measure. Now the compiler knows.