First, flimsy pass at GUI cursor j/k movement
It has at least these flaws:
 + l works like j and h works like k
 + Many cases of lacking bounds checking can blow it sky high once you
   reach the edge of the screen
 + As a corollary of the boundary cases, it does not scroll at the edges
 + Redraws the whole screen when it could just redraw two cells

Notice that I managed to destroy the Motion tag interface. Thinking
deeply about Haskell has made me a better person and shown me the
disadvantages of subtyping. I am now trying to move this F# code to be
more and more purely functional and to be as Haskell-esque as possible
because I believe it will lead me to write better code.

One problem with tag interfaces is they do not compose. For example:

    type BufferCommandMessage =
        ...
        interface BufferMessage
        interface CommandMessage

There is a strong divorce between looking at BufferCommandMessage as a
CommandMessage and looking at it as a BufferMEssage. You can't say,
"anything that is a BufferMessage and a CommandMessage" generically; you
either have to say, "anything that is a BufferMessage" generically, or
"anything that is a CommandMessage" generically, or "a
BufferCommandMessage" concretely, but you can't marry the two interfaces
up and compose them generically.

But what if I shifted my OO tag-interface mentality to an FP
perspective, embraced algebraic data types that much more, and made tag
wrappers instead? Like newtype in Haskell?

    type Message<'a> = Message of 'a
    type BufferMessage<'a> = BufferMessage of 'a
    type CommandMessage<'a> = CommandMessage of 'a

That should compose. Then you may have an isomorphism problem, I
suppose: if you want to refer to a buffer command message generically,
is it a BufferMessage<CommandMessage<'a>>, or
CommandMessage<BufferMessage<'a>>? So I obviously haven't thought this
out all the way, but... algebraic data types do compose; that is a large
part of their beauty.