Get the window and buffer cursors in sync
This should fix the bug described before.

However, it seems like a smell that we are always handling both row and
line. Units of measure are helping me to keep from clobbering myself
with this, but it's still messy. I should look at separating that out
before submitting a pull request.