Create installer for Void!
FAKE allowed me to get up and running with WiX real quickly. That's a
win!

However if you run through it there's a bunch of Lorem Ipsum and
whatnot...

Also, it installs all the test code as well, which it should not.

Also not sure it modifies the path at all.