Remove test binaries from MSI installer
I want to make sure that the tests run against the actual code that is
going to get deployed, so I don't want to build to two separate
directories, release and test, or something. However, there are obvious
potential flaws with the way I've done it of building everything to one
directory, running the tests and then actively removing the things I
don't want (having to maintain a list of things of patterns of things to
remove, potentially accidentally (though unlikely) removing something
that shouldn't be if the projects change, not removing something that I
should, not being able to re-run the binaries to repeat test
failures...) I suppose the other option is to build just the stuff I
want, build the install, then build the rest and run the tests. But I
definitely don't want to build two different ways. I seem to remember
there were issues with building projects individually but I don't
remember what they were (performance?). Anyhow, I'm going to try this
for now and see how it goes.